add logica

+++ components/MaintenanceActivity.tsx
@@ -6,31 +6,16 @@ import {
   Box,
   Button,
   Modal,
-  TextField,
-  FormControl,
-  Select,
-  MenuItem,
-  InputLabel,
 } from "@mui/material";
-import { SelectChangeEvent } from "@mui/material/Select";
 import { useRouter } from "next/router";
-import InputMask from "react-input-mask";
-import {
-  Activity,
-  fetchBlocks,
-  getActivityHistory,
-} from "@/services/firebaseService";
+import { Activity, fetchBlocks } from "@/services/firebaseService";
 import ActivityStatus from "@/components/layout/ActivityStatus";
-import HelpActivity from "@/utils/HelpActivity";
-import { DatePicker } from "@mui/x-date-pickers/DatePicker"; // Import do DatePicker
-import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider"; // Import do LocalizationProvider
-import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs"; // Adaptador para Day.js
-import dayjs, { Dayjs } from "dayjs"; // Import para trabalhar com datas
+import EditActivityModal from "./EditActivityModal"; // Import do modal EditActivityModal
 
 interface MaintenanceActivityProps {
   activity: Activity;
   onUpdate: (updatedActivity: Activity) => void;
-  onRemove: (activityId: number) => void;
+  onRemove?: (activityId: number) => void;
   removeValid: boolean;
   titleUpdate: string;
 }
@@ -42,19 +27,13 @@ const MaintenanceActivity: React.FC<MaintenanceActivityProps> = ({
   removeValid,
   titleUpdate,
 }) => {
-  const [open, setOpen] = useState(false);
   const [removeOpen, setRemoveOpen] = useState(false);
   const [historyOpen, setHistoryOpen] = useState(false);
-  const [editedActivity, setEditedActivity] = useState<Activity | null>(null);
+  const [selectedActivity, setSelectedActivity] = useState<Activity | null>(
+    null
+  );
+  const [modalOpen, setModalOpen] = useState(false);
   const [blocks, setBlocks] = useState<{ id: string; name: string }[]>([]);
-  const [selectedBlocks, setSelectedBlocks] = useState<string[]>([]);
-  const [history, setHistory] = useState<any[]>([]);
-  const [activityRegular, setActivityRegular] = useState<boolean>(false);
-  const [periodicityOptions, setPeriodicityOptions] = useState<string[]>([]);
-  const [formattedLastMaintenance, setFormattedLastMaintenance] = useState<
-    string | null
-  >(null);
-  const [formattedDueDate, setFormattedDueDate] = useState<string | null>(null);
 
   const router = useRouter();
 
@@ -63,42 +42,17 @@ const MaintenanceActivity: React.FC<MaintenanceActivityProps> = ({
       const fetchedBlocks = await fetchBlocks();
       setBlocks(fetchedBlocks || []);
     };
-    fetchPeriodicityOptions();
     loadBlocks();
   }, []);
 
-  useEffect(() => {
-    const fetchFormattedDates = async () => {
-      if (activity.data) {
-        const lastMaintenanceDate = await HelpActivity.formatDate(
-          activity.data
-        );
-        setFormattedLastMaintenance(lastMaintenanceDate);
-      }
-
-      const dueDate = await HelpActivity.formatDateToDDMMYYYY(activity);
-      setFormattedDueDate(dueDate);
-    };
-
-    fetchFormattedDates();
-  }, [activity]);
-
-  const handleOpen = () => {
-    if (!activity || Object.keys(activity).length === 0) {
-      console.error("Activity is undefined or empty");
-      return; // Sai da função se 'activity' for indefinido ou vazio
-    }
-    // Verifica se 'activity' está corretamente definido antes de atualizar os estados
-    setEditedActivity(activity);
-    setActivityRegular(activity.activityRegular || false);
-    setSelectedBlocks(activity.blocoIDs || []);
-    setOpen(true);
+  const handleOpenModal = (activity: Activity) => {
+    setSelectedActivity(activity);
+    setModalOpen(true);
   };
 
-  const handleClose = () => {
-    setOpen(false);
-    setEditedActivity(null);
-    setSelectedBlocks([]);
+  const handleCloseModal = () => {
+    setSelectedActivity(null);
+    setModalOpen(false);
   };
 
   const handleRemoveOpen = () => setRemoveOpen(true);
@@ -108,66 +62,10 @@ const MaintenanceActivity: React.FC<MaintenanceActivityProps> = ({
     router.push(`/activity/${activity.id}`);
   };
 
-  const handleHistoryClose = () => setHistoryOpen(false);
-
-  const handleSelectChange = useCallback(
-    (event: SelectChangeEvent<string[]>) => {
-      const value = Array.isArray(event.target.value) ? event.target.value : [];
-      setSelectedBlocks(value);
-    },
-    []
-  );
-
-  const handleSelectChangePeriodicidade = useCallback(
-    (event: SelectChangeEvent<string>) => {
-      const value = event.target.value;
-      setEditedActivity((prevActivity) =>
-        prevActivity ? { ...prevActivity, Periodicidade: value } : null
-      );
-    },
-    []
-  );
-
-  const fetchPeriodicityOptions = async () => {
-    try {
-      const response = await fetch("/periodicidades/periodicidade.json");
-      const result = await response.json();
-
-      const options = result.map(
-        (item: { descricao: string }) => item.descricao
-      );
-      setPeriodicityOptions(options);
-    } catch (error) {
-      console.error("Erro ao buscar as opções de periodicidade:", error);
-    }
-  };
-
-  const handleChange = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value, type, checked } = e.target;
-    setEditedActivity((prevActivity) =>
-      prevActivity
-        ? { ...prevActivity, [name]: type === "checkbox" ? checked : value }
-        : null
-    );
-  }, []);
-
-  const handleSave = () => {
-    if (editedActivity) {
-      const finalBlocks =
-        selectedBlocks.length === 0 ? blocks.map((b) => b.id) : selectedBlocks;
-
-      const updatedActivity = {
-        ...editedActivity,
-        blocoIDs: finalBlocks,
-        activityRegular,
-      };
-      onUpdate(updatedActivity);
-      handleClose();
-    }
-  };
-
   const handleRemove = () => {
-    onRemove(activity.id);
+    if (onRemove) {
+      onRemove(activity.id);
+    }
     handleRemoveClose();
   };
 
@@ -195,10 +93,10 @@ const MaintenanceActivity: React.FC<MaintenanceActivityProps> = ({
             <strong>Periodicidade:</strong> {activity.Periodicidade}
           </Typography>
 
-          {Array.isArray(selectedBlocks) && selectedBlocks.length > 0 && (
+          {Array.isArray(activity.blocoIDs) && activity.blocoIDs.length > 0 && (
             <Typography variant="body2">
               <strong>Blocos:</strong>{" "}
-              {selectedBlocks
+              {activity.blocoIDs
                 .map(
                   (blockId) =>
                     blocks.find((block) => block.id === blockId)?.name
@@ -211,10 +109,10 @@ const MaintenanceActivity: React.FC<MaintenanceActivityProps> = ({
           {activity.data && (
             <>
               <Typography variant="body2">
-                Última manutenção: {formattedLastMaintenance || "Carregando..."}
+                Última manutenção: {activity.data || "Carregando..."}
               </Typography>
               <Typography variant="body2">
-                Vencimento: {formattedDueDate || "Carregando..."}
+                Próxima manutenção: {activity.data || "Carregando..."}
               </Typography>
             </>
           )}
@@ -228,7 +126,11 @@ const MaintenanceActivity: React.FC<MaintenanceActivityProps> = ({
               gap: 2,
             }}
           >
-            <Button variant="contained" color="primary" onClick={handleOpen}>
+            <Button
+              variant="contained"
+              color="primary"
+              onClick={() => handleOpenModal(activity)}
+            >
               {titleUpdate}
             </Button>
             {activity.data && (
@@ -253,152 +155,19 @@ const MaintenanceActivity: React.FC<MaintenanceActivityProps> = ({
         </CardContent>
       </Card>
 
-      <Modal open={open} onClose={handleClose}>
-        <Box
-          sx={{
-            position: "absolute",
-            top: "50%",
-            left: "50%",
-            transform: "translate(-50%, -50%)",
-            width: 400,
-            bgcolor: "background.paper",
-            border: "2px solid #000",
-            boxShadow: 24,
-            p: 4,
-          }}
-        >
-          <Typography variant="h6" component="h2">
-            Editar Atividade
-          </Typography>
-          <TextField
-            fullWidth
-            margin="normal"
-            label="Título"
-            name="titulo"
-            value={editedActivity?.titulo || ""}
-            onChange={handleChange}
-          />
-          <TextField
-            fullWidth
-            margin="normal"
-            label="Atividade"
-            name="atividade"
-            value={editedActivity?.atividade || ""}
-            onChange={handleChange}
-          />
-          <TextField
-            fullWidth
-            margin="normal"
-            label="Responsável"
-            name="responsavel"
-            value={editedActivity?.responsavel || ""}
-            onChange={handleChange}
-            disabled={true}
-          />
-          <FormControl
-            fullWidth
-            margin="normal"
-            required={
-              editedActivity?.Periodicidade ===
-              "Conforme indicação dos fornecedores"
-            }
-          >
-            <InputLabel>Periodicidade</InputLabel>
-            <Select
-              label="Periodicidade"
-              name="Periodicidade"
-              value={editedActivity?.Periodicidade || ""}
-              onChange={handleSelectChangePeriodicidade}
-              disabled={
-                editedActivity?.Periodicidade !==
-                "Conforme indicação dos fornecedores"
-              }
-            >
-              {periodicityOptions
-                .filter(
-                  (option) =>
-                    option !== "Conforme indicação dos fornecedores" &&
-                    option !==
-                      "A cada 5 anos para edifícios de até 10 anos de entrega, A cada 3 anos para edifícios entre 11 a 30 anos de entrega, A cada ano para edifícios com mais de 30 anos de entrega"
-                )
-                .map((option, index) => (
-                  <MenuItem key={index} value={option}>
-                    {option}
-                  </MenuItem>
-                ))}
-            </Select>
-          </FormControl>
+      {/* Modal de Edição */}
+      {selectedActivity && (
+        <EditActivityModal
+          open={modalOpen}
+          activity={selectedActivity}
+          onClose={handleCloseModal}
+          onActivityUpdated={() => onUpdate(selectedActivity)}
+          isEdit={true}
+          showNotApplicable={true}
+        />
+      )}
 
-          {blocks.length > 0 && (
-            <FormControl fullWidth margin="normal">
-              <InputLabel>Bloco</InputLabel>
-              <Select
-                multiple
-                label="Bloco"
-                value={Array.isArray(selectedBlocks) ? selectedBlocks : []}
-                onChange={handleSelectChange}
-                renderValue={(selected) =>
-                  Array.isArray(selected)
-                    ? selected
-                        .map(
-                          (selectedId) =>
-                            blocks.find((block) => block.id === selectedId)
-                              ?.name
-                        )
-                        .join(", ")
-                    : ""
-                }
-              >
-                {blocks.map((block) => (
-                  <MenuItem key={block.id} value={block.id}>
-                    {block.name}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-          )}
-
-          {editedActivity?.Periodicidade !== "Não aplicável" ? (
-            <LocalizationProvider dateAdapter={AdapterDayjs}>
-              <DatePicker
-                label="Data"
-                value={dayjs(editedActivity?.data)} // Converte a string de data para o formato dayjs
-                onChange={(newDate) =>
-                  setEditedActivity((prevActivity) =>
-                    prevActivity
-                      ? { ...prevActivity, data: newDate?.format("DD/MM/YYYY") }
-                      : null
-                  )
-                }
-                slotProps={{
-                  textField: { fullWidth: true }, // Para renderizar o TextField com largura total
-                }}
-                format="DD/MM/YYYY" // Formato desejado
-              />
-            </LocalizationProvider>
-          ) : (
-            <Box sx={{ display: "flex", justifyContent: "flex-start", mt: 2 }}>
-              <Button
-                variant="contained"
-                color={activityRegular ? "success" : "primary"}
-                onClick={() => setActivityRegular(!activityRegular)}
-              >
-                {activityRegular ? "Feito" : "Marcar como Feito"}
-              </Button>
-            </Box>
-          )}
-
-          <Box sx={{ display: "flex", justifyContent: "space-between", mt: 2 }}>
-            <Button variant="contained" color="secondary" onClick={handleClose}>
-              Cancelar
-            </Button>
-            <Button variant="contained" color="primary" onClick={handleSave}>
-              Salvar
-            </Button>
-          </Box>
-        </Box>
-      </Modal>
-
+      {/* Modal de Remoção */}
       <Modal open={removeOpen} onClose={handleRemoveClose}>
         <Box
           sx={{
@@ -429,49 +198,6 @@ const MaintenanceActivity: React.FC<MaintenanceActivityProps> = ({
             </Button>
             <Button variant="contained" color="primary" onClick={handleRemove}>
               Remover
-            </Button>
-          </Box>
-        </Box>
-      </Modal>
-
-      <Modal open={historyOpen} onClose={handleHistoryClose}>
-        <Box
-          sx={{
-            position: "absolute",
-            top: "50%",
-            left: "50%",
-            transform: "translate(-50%, -50%)",
-            width: 400,
-            bgcolor: "background.paper",
-            border: "2px solid #000",
-            boxShadow: 24,
-            p: 4,
-          }}
-        >
-          <Typography variant="h6" component="h2">
-            Histórico de Alterações
-          </Typography>
-          {history.length > 0 ? (
-            history.map((entry, index) => (
-              <Box key={index} sx={{ mb: 2 }}>
-                <Typography variant="body2">
-                  <strong>Data:</strong>{" "}
-                  {new Date(entry.timestamp.seconds * 1000).toLocaleString()}
-                </Typography>
-              </Box>
-            ))
-          ) : (
-            <Typography variant="body2">
-              Nenhum histórico disponível.
-            </Typography>
-          )}
-          <Box sx={{ display: "flex", justifyContent: "flex-end", mt: 2 }}>
-            <Button
-              variant="contained"
-              color="primary"
-              onClick={handleHistoryClose}
-            >
-              Fechar
             </Button>
           </Box>
         </Box>
